Proper indirect buffer object bindings added in places where indirect drawing is used. Indices type changed to GL_UNSIGNED_BYTE/SHORT where the number of indices are small enough

+++ src/model/Model.cpp
@@ -49,7 +49,7 @@ void Model::loadInstances(glm::mat4 *models, unsigned int numModels)
 Mesh Model::processMesh(aiMesh *mesh, const aiScene* scene)
 {
   std::vector<Vertex> vertices;
-  std::vector<unsigned int> indices;
+  std::vector<GLushort> indices;
   std::vector<Texture> textures;
 
   for (unsigned int i = 0; i < mesh->mNumVertices; i++)
